Made some changes to file upload to keep routers clean. Need to extract file parsing and s3 upload into a middleware.
