+++ api/upload/uploadModel.js
@@ -1,0 +1,7 @@
+const db = require('../../data/db-config');
+
+const addAvatar = (URL) => db('Avatars').insert(URL).returning('*');
+
+module.exports = {
+  addAvatar,
+};
